tRY cATCH eXCEPTIONS ADDED

+++ Ubrania_Nowy/Ubrania_Nowy/ERP.cs
@@ -12,8 +12,8 @@ namespace Ubrania_Nowy
         public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
-        public int Tel { get; set; }
-        public int Pesel { get; set; }
+        public double Tel { get; set; }
+        public double Pesel { get; set; }
         public DateTime Begin { get; set; }        
         public DateTime End { get; set; }
         public IList<Cloth> Clothes { get; set; }
